pyproject.toml
[project]
name = "livebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.13",
    "aiosu>=2.3.5",
    "nonebot-adapter-bilibili-live>=0.2.4",
    "nonebot-plugin-danmaku",
    "nonebot2[aiohttp,fastapi]>=2.4.2",
    "pydle>=1.1.0",
    "simpleobsws>=1.4.3",
]

[tool.nonebot]
plugin_dirs = ["src/plugins"]

[tool.nonebot.plugins]
"nonebot-plugin-danmaku" = ["nonebot_plugin_danmaku"]

[tool.nonebot.adapters]
nonebot-adapter-bilibili-live = [{name = "B站直播间", module_name = "nonebot.adapters.bilibili_live"}]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
  "F",     # Pyflakes
  "W",     # pycodestyle warnings
  "E",     # pycodestyle errors
  "I",     # isort
  "UP",    # pyupgrade
  "ASYNC", # flake8-async
  "C4",    # flake8-comprehensions
  "T10",   # flake8-debugger
  "PYI",   # flake8-pyi
  "PT",    # flake8-pytest-style
  "Q",     # flake8-quotes
  "TID",   # flake8-tidy-imports
  "RUF",   # Ruff-specific rules
  "FAST",  # FastAPI
  "YTT",   # flake8-2020
  "S",     # flake8-bandit
  "INP",   # flake8-no-pep420
  "SIM",   # flake8-simplify
  "ARG",   # flake8-unused-arguments
  "PTH",   # flake8-use-pathlib
  "N",     # pep8-naming
  "FURB"   # refurb
]
ignore = [
  "E402",   # module-import-not-at-top-of-file
  "UP037",  # quoted-annotation
  "RUF001", # ambiguous-unicode-character-string
  "RUF002", # ambiguous-unicode-character-docstring
  "RUF003", # ambiguous-unicode-character-comment
  "S101",   # assert
  "S311",   # suspicious-non-cryptographic-random-usage
]


[tool.ruff.lint.isort]
force-sort-within-sections = true
force-wrap-aliases = true
combine-as-imports = true
order-by-type = true
relative-imports-order = "closest-to-furthest"
known-first-party = ["nonebot_adapter_osu"]
required-imports = ["from __future__ import annotations"]
extra-standard-library = ["typing_extensions"]
section-order = ["future", "standard-library", "first-party", "local-folder", "third-party"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.ruff.lint.extend-per-file-ignores]
"src/plugins/**/*.py" = ["INP001"]

[tool.pyright]
pythonVersion = "3.12"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }

typeCheckingMode = "standard"
disableBytesTypePromotions = true
reportIncompatibleMethodOverride = false
reportIncompatibleVariableOverride = false

[tool.uv.workspace]
members = [
    "src/packages/nonebot-plugin-danmaku",
]

[tool.uv.sources]
nonebot-plugin-danmaku = { workspace = true }

[dependency-groups]
dev = [
    "nb-cli>=1.7.0",
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "ruff>=0.12.1",
]

